Make CoreAnnotationNames internal
Fixes #17159
